+++ src/games/abande.ts
@@ -55,7 +55,8 @@ export class AbandeGame extends GameBase {
                 group: "board",
                 description: "A 37-space hexagonal board."
             }
-        ]
+        ],
+        flags: ["limited-pieces", "scores"]
     };
 
     public numplayers: number = 2;
@@ -475,6 +476,10 @@ export class AbandeGame extends GameBase {
 
     protected getMoveList(): any[] {
         return this.getMovesAndResults(["move", "place"]);
+    }
+
+    public getPlayerPieces(player: number): number {
+        return this.pieces[player - 1];
     }
 
     public getPlayerScore(player: number): number {