+++ src/games/mvolcano.ts
@@ -68,7 +68,8 @@ export class MvolcanoGame extends GameBase {
                 name: "Kristin Looney",
                 urls: ["http://www.wunderland.com/WTS/Kristin/Kristin.html"]
             }
-        ]
+        ],
+        flags: ["shared-pieces", "scores", "stacking-expanding"]
     };
     public static coords2algebraic(x: number, y: number): string {
         return GameBase.coords2algebraic(x, y, 6);