+++ src/games/volcano.ts
@@ -87,7 +87,8 @@ export class VolcanoGame extends GameBase {
                 name: "Kristin Looney",
                 urls: ["http://www.wunderland.com/WTS/Kristin/Kristin.html"]
             }
-        ]
+        ],
+        flags: ["shared-pieces", "stacking-expanding"]
     };
     public static coords2algebraic(x: number, y: number): string {
         return GameBase.coords2algebraic(x, y, 5);